refactoring of redux, reducers and actions
