+++ src/actions/types.js
@@ -1,0 +1,4 @@
+export const ActionType = {
+  CHANGE_CITY: `CHANGE_CITY`,
+  GET_OFFERS: `GET_OFFERS`
+};
