+++ src/index.js
@@ -1,17 +1,15 @@
 import App from './components/app/app.jsx';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {Provider} from 'react-redux';
 import {offers} from './mocks/offers.js';
-import {cities} from './mocks/cities.js';
+import {store} from './store';
 
-const init = () => {
-  ReactDOM.render(
+ReactDOM.render(
+    <Provider store={store}>
       <App
         offers={offers}
-        cities={cities}
-      />,
-      document.querySelector(`#root`)
-  );
-};
-
-init();
+      />
+    </Provider>,
+    document.querySelector(`#root`)
+);
