+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import {combineReducers} from 'redux';
+import offersReducer from './offers-reducer';
+
+export default combineReducers({
+  offers: offersReducer
+});
