+++ src/reducers/offers-reducer.test.js
@@ -1,39 +1,40 @@
-import {reducer, ActionType} from './reducer';
-import {offers} from './mocks/offers';
-import {cities, offers as mockOffers} from './test-data';
+import reducer from './offers-reducer';
+import {offers} from '../mocks/offers';
+import {cities, offers as mockOffers} from '../test-data';
+import {ActionType} from '../actions/types';
 
 describe(`Reducer works correctly`, () => {
   it(`Reducer has to return initial state if new state is undefined`, () => {
     expect(reducer(undefined, {}))
     .toEqual({
       activeCity: `Amsterdam`,
-      activeOffers: offers
+      offers
     });
   });
 
   it(`Reducer should return new state with new city`, () => {
     expect(reducer({
       activeCity: `Amsterdam`,
-      activeOffers: offers
+      offers
     }, {
       type: ActionType.CHANGE_CITY,
       city: cities[0].name
     })).toEqual({
       activeCity: `Gomel`,
-      activeOffers: offers
+      offers
     });
   });
 
   it(`Reducer should return new state with new offers`, () => {
     expect(reducer({
       activeCity: `Amsterdam`,
-      activeOffers: offers
+      offers
     }, {
       type: ActionType.GET_OFFERS,
       offers: mockOffers
     })).toEqual({
       activeCity: `Amsterdam`,
-      activeOffers: mockOffers
+      offers: mockOffers
     });
   });
 });
