+++ src/store.js
@@ -1,0 +1,7 @@
+import {createStore} from 'redux';
+import reducer from './reducers/offers-reducer';
+
+export const store = createStore(
+    reducer,
+    window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : (f) => f
+);
